hopefully cross platform to_string for display of active float value

+++ src/dtpulse.cpp
@@ -426,7 +426,7 @@ std::string AbsoluteSequence::getWorkingStepDisplay() {
     return str;
   }
   else {
-    return std::to_string(exactFloats[stepIndex - 52]);
+		return std::to_string((long double) exactFloats[stepIndex - 52]);
   }
 }
